Move CachingDirectoryLister to fs package
