+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/fs/CachingDirectoryListerModule.java
@@ -11,11 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive;
+package io.trino.plugin.hive.fs;
 
 import com.google.inject.Binder;
 import com.google.inject.Module;
 import com.google.inject.Scopes;
+import io.trino.plugin.hive.TableInvalidationCallback;
 
 import java.util.Optional;
